+++ src/pacioli/src/main/java/pacioli/lsp/PacioliLanguageServer.java
@@ -2,6 +2,8 @@ package pacioli.lsp;
 
 import java.util.concurrent.CompletableFuture;
 
+import org.eclipse.lsp4j.CompletionOptions;
+import org.eclipse.lsp4j.DiagnosticRegistrationOptions;
 import org.eclipse.lsp4j.InitializeParams;
 import org.eclipse.lsp4j.InitializeResult;
 import org.eclipse.lsp4j.MessageParams;
@@ -67,6 +69,9 @@ public class PacioliLanguageServer implements LanguageServer, LanguageClientAwar
         final InitializeResult response = new InitializeResult(new ServerCapabilities());
         // Set the document synchronization capabilities to full.
         response.getCapabilities().setTextDocumentSync(TextDocumentSyncKind.Full);
+        response.getCapabilities().setCompletionProvider(new CompletionOptions());
+        response.getCapabilities().setDiagnosticProvider(new DiagnosticRegistrationOptions());
+
         var clientCapabilities = params.getCapabilities();
 
         /*